Fix uppercase function
Fix the `uppercase` function to not depend on undefined behavior.

`sPtr` was used twice in a statement, of which once in a in-place
increment - in that case the sequence depends on C compiler, and for
many compilers this resulted in removing the first character.

Explicitly delete the first character of the result in `write_fd`,
the only callsite.