fix column header display
The centering of the column header is done by padding with spaces on the
left. This padding is at least one space. This breaks the header display
when a column width is set to 2 and the header is a double-letter as the
second letter overflows the field and gets overwritten.

Rework the padding computation so that double-letter headings are always
properly centered, including in the extreme case with a width of 2 where
no padding at all should be applied.

This also properly aligns single-letter headers with centered single-letter
cell content when the field width is odd.