de-obscurify ui_show_content()
There was a bug with cursor positioning for the "show-cursor" option.

The problem was in ui_show_content(). The cursor was always one row
too high.

This function was very hard to understand and debug. So I simplified
the loop construction to make it clearer and more logical.

And in doing so, the original bug just disappeared!

I therefore applied the same cleanup to ui_show_sc_col_headings() and
ui_show_sc_row_headings().

The location of the rightward and downward frozen block is already found
in calc_offscr_sc_cols() and calc_offscr_sc_rows() so this is remembered
in the first_frozen_after_cols and first_frozen_after_rows global
variables rather than being recomputed all the time.