adjust row/col count on edge cases
If one follows the sequence in issue #578 (which is now fixed) the
selected cell ends up on row 33. But moving to row 34 will leave
4 empty lines at the bottom of the screen.

Fix that by attempting to fill up the screen when there is still room
after counting rows/cols in the backward case.

Also fix the returned "last" row/col in calc_mobile_rows() and
calc_mobile_cols() to always be the last one to fit on screen.
even in the backward case.