unspecial: LaTex: no quotes, escape underscore
For writing a LaTex file, don't put quotes around the cell
when the special characters already get escaped.
In LaTex, the doublequote is not a special character, so
they are printed literally in the pdf output.

The underscore is a special character in LaTex, so add
that to the list of characters to escape.