Attempt to fix the Lua version situation
Drawing heavily on #589 by @dlbeer, attempt to fix the utter
nonsense pkg-config and all various distributors brought on the
world by

- Normalising 'lua-X.Y' and 'luaXY' notations to mean the same thing
- Handle both notations existing on the system correctly
- Picking the numerically highest version

There is also an option to override this autodetection mechanism by
setting the `LUA_PKGNAME` environment variable prior to kicking a
build off, resulting in the specified Lua package to be used.

Tested to be working on FreeBSD 13.1, Ubuntu 22.04 and Debian 11.